+++ Common/PresentationLayer/Presenters/TranslatedMessagesPresenter.vb
@@ -25,7 +25,7 @@ Namespace PresentationLayer.Presenters
             Ea.SubscribeEvent(Me)
         End Sub
 
-        'Public Overrides Sub Display(messageIdNo As Int32)
+        'Public Overrides Sub Display(messageIdNo As Int16)
         '    Dim idNoTm As Integer
         '    idNoTm = GetRecordFieldWithKey(messageIdNo, "TranslatedMessages", "messageIdNo", "IdNo")
         '    Dim modelData As New TranslatedMessagesModel
