+++ IspDataDb/dbo/Tables/AccountTypes.sql
@@ -1,9 +1,11 @@
 ﻿CREATE TABLE [dbo].[AccountTypes] (
-    [IdNo]         INT          IDENTITY (1, 1) NOT NULL,
+    [IdNo]         SMALLINT     IDENTITY (1, 1) NOT NULL,
     [AccountIdNo]  INT          NOT NULL,
     [AccountTypes] VARCHAR (50) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     CONSTRAINT [PK_AccountTypesIdNo] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
 
 
 GO
