+++ IspDataDb/dbo/Tables/CashCode.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[CashCode] (
-    [IdNo]                      INT            IDENTITY (1, 1) NOT NULL,
+    [IdNo]                      TINYINT        IDENTITY (1, 1) NOT NULL,
     [CashCode]                  CHAR (1)       COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
     [CashName]                  NVARCHAR (30)  COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
     [CashNameAra]               NVARCHAR (30)  COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
@@ -11,4 +11,6 @@
     [Notes]                     NVARCHAR (255) COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
     CONSTRAINT [PK_CashCode] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
