+++ IspDataDb/dbo/Tables/EmployeeAbsencesLeave.sql
@@ -1,11 +1,13 @@
 ﻿CREATE TABLE [dbo].[EmployeeAbsencesLeave] (
     [IdNo]             INT      NULL,
-    [EmployeeIdNo]     SMALLINT NULL,
+    [EmployeeIdNo]     INT      NULL,
     [LeaveAbsenceIdNo] SMALLINT NULL,
     [StartDate]        DATE     NULL,
     [EndDate]          DATE     NULL,
     [Approved]         BIT      NULL
 );
+
+
