+++ IspDataDb/dbo/Tables/PayInformation.sql
@@ -1,10 +1,12 @@
 ﻿CREATE TABLE [dbo].[PayInformation] (
     [IdNo]             INT      NULL,
-    [EmployeeIdNo]     SMALLINT NULL,
+    [EmployeeIdNo]     INT      NULL,
     [FrequencyOfPay]   CHAR (1) NULL,
     [SalariedOrHourly] CHAR (1) NULL,
     [Rate]             MONEY    NULL
 );
+
+
